add a test that counts the output to see if optimizations are culling results
